Merged PR 543525: Fixed directed broadcast functionality after VlanMgr changes
After bootup, directed broadcast daemon retries to fetch the interface data every 5 seconds if not populated. This update is within a timer logic and a fix is made to update the dictionary if it is empty.

Why was it not seen before?
With the introduction of VlanMgr, the interface table for "VLAN_INTERFACE" is not populated by the time directed_bcast daemon is started. Previously, when the daemon started, the table was already populated.

Logs:
Nov 30 22:45:56.526277 str-z9100-acs-2 ERR directed_bcastd: Failed to instantiate Interface object for interface Vlan1000 ([Errno 19] No such device)