Merged PR 2799725: Apply AMD patch for all image types except NBI.
Apply AMD patch for all image types except NBI.
This is because, with microcode prepended to initrd.img, the build is unable to handle.

Details on this patch as obtained from Arista is copy/pasted below.

From: Samuel Angebault staphylo@arista.com
Sent: Wednesday, April 8, 2020 9:52 AM
To: Renuka Manavalan Renuka.Manavalan@microsoft.com; sonic-ext-support sonic-ext-support@arista.com
Cc: Guohan Lu gulv@microsoft.com; Ying Xie Ying.Xie@microsoft.com; Joe LeVeque jolevequ@microsoft.com; Vasiliy Khoruzhick vasilykh@arista.com
Subject: [EXTERNAL] Re: Looking for help

+sonic-ext-support

Hi Renuka,

What you observe is a property of initrds and how the microcode early update is supposed to work (amd and intel)
It is possible to concatenate as many initrd as you want and the linux kernel will be able to load them all to build the final memory filesystem.
The debian initrd is still there compressed as a gz file but the amd64_microcode hook generate another uncompressed initrd with the firmwares and prepend this early initramfs to the main one.

Unfortunately the cpio tool is not smart enough to deal with such initrds and only display one of them.
There are alternative tools like skipinitrd that can manage concatenated initrds (parts of dracut).

The fix we provided is almost the same one that we planned on upstreaming once AMD publishes the microcode in linux-firmware.
The only difference is that we would get rid of the base64 encoded microcode.

When making the fix I made sure that nothing else was using cpio on the initrd to avoid such an issue.
It seems that the Cisco code that you mention does not exist in the open source world which is why I couldn't have anticipated this fallback.

The way I see it there are only 2 solutions

rewrite the hook so it puts the microcode into the main cpio
However I don't believe this would work since I believe the kernel has expectations about having the microcode in an uncompressed initrd for early microcode updates.
This also applies to the intel-microcode package which has a really similar initramfs hook for its microcode.
This would take some time to rewrite and validate.

Only install amd64-microcode for Arista SWI images
The caveat is that build_debian.sh is not meant to be aware of which final packaging mechanism will be used.
Usually only build_image.sh knows about the final target.

Thanks,