Merged PR 3608822: [Monit] Fix the syntax error of Monit configuration file of ACMS container.
Currently Monit will periodically run a python script to check whether a critical process in a container is running or not. Although the Monit configuration file of ACMS container was changed from `check process ...` to `check program ...` according to the new format, we also need change the syntax of alerting statement from `if does not exist ...` to `if status != 0`.

Signed-off-by: Yong Zhao <yozhao@microsoft.com>