Merged PR 4446401: Updated platform_modules/cisco_sfp for 3164
Updated platform_modules/cisco_sfp for 3164

1. Fixing the error "QSFP read NACK error" in platform_modules.
2. Toggling qsfp reset with a 1 sec delay in unreset API in sfp plugin. This gets the interface back up after QSFP reseat

Verified that the interfaces comes up after reseating the optics in case of  (AOI, Finisar, o-net, wavesplitter).
The eeprom data corruption was not seen after reboot, but have seen this corrupt data with AOI QSFP after reseating - as it takes a while to make the interface UP. This might need some more tuning.