Merged PR 5374386: [vnet-monitor] sniff on all active ifaces
help(sniff) of 2.4.4 says that sniff all interfaces if iface=null. However, one of the online document says that sniff only conf.iface if iface=null. I confirmed that the later one is correct. And sniffing on a dead interface can cause exception. So that only active interfaces are sniffed on.