Merged PR 5902982: Skip AN setting in minigraph for SN2700
Speed advertisement attribute SAI_PORT_ATTR_ADVERTISED_SPEED is NOT supported by Mellanox SAI in 201911. When AN is present in minigraph, the syncd crashes due to above attribute being not supported, but being set by Orchagent when AN is enabled. To avoid this crash, we skip setting of AN via minigraph for SN700 platform. Please NOTE, by default AN is enabled in 201911 for SN700 platform.