Merged PR 7338307: [202211][bgp-cfgd] BGP template for OLT mgmt ToR
Why I did it
For mgmt ToR which is connected to OLT, need to have bgp allow-prefix list to control the routes advertisement between OLT and mgmt ToR.
How I did it
To add new template for this Scenario, which is based on common mgmt ToR BGP template.
below is the doc for detailed design.
https://microsoft.sharepoint.com/:w:/t/Aznet/EUXF5cffHVtDspT4W0o3ni8Bax5uGLwe5_hRGQnERDYrbg?e=rtmE9c
The changes in managers_allow_list.py is a common one and PR in public repo.
https://github.com/Azure/sonic-buildimage/pull/11586
How to verify it
1. Unit Test
Add new unit test cases for new templates.
2. Integration Test
To simulate the Scenario in physical TB, and verify the run bgp configuration.
WIP, to verify with Optical team.

Others
Need to port back to 202012 and 202205 branch.

Cherry picked from !6446272